+++ shell/src/main/java/alluxio/cli/fs/command/DistributedLoadCommand.java
@@ -15,18 +15,9 @@ import alluxio.AlluxioURI;
 import alluxio.annotation.PublicApi;
 import alluxio.cli.CommandUtils;
 import alluxio.cli.fs.FileSystemShellUtils;
-import alluxio.cli.fs.command.job.JobAttempt;
 import alluxio.client.file.FileSystemContext;
-import alluxio.client.file.URIStatus;
-import alluxio.client.job.JobMasterClient;
 import alluxio.exception.AlluxioException;
 import alluxio.exception.status.InvalidArgumentException;
-import alluxio.grpc.ListStatusPOptions;
-import alluxio.job.JobConfig;
-import alluxio.job.plan.load.LoadConfig;
-import alluxio.job.wire.JobInfo;
-import alluxio.retry.CountingRetry;
-import alluxio.retry.RetryPolicy;
 
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
@@ -109,108 +100,12 @@ public final class DistributedLoadCommand extends AbstractDistributedJobCommand 
     String[] args = cl.getArgs();
     AlluxioURI path = new AlluxioURI(args[0]);
     int replication = FileSystemShellUtils.getIntArg(cl, REPLICATION_OPTION, DEFAULT_REPLICATION);
-    distributedLoad(path, replication);
+    DistributedLoadUtils.distributedLoad(this, path, replication);
     return 0;
   }
 
   @Override
   public void close() throws IOException {
     mClient.close();
-  }
-
-  /**
-   * Creates a new job to load a file in Alluxio space, makes it resident in memory.
-   *
-   * @param filePath The {@link AlluxioURI} path to load into Alluxio memory
-   * @param replication The replication of file to load into Alluxio memory
-   */
-  private LoadJobAttempt newJob(AlluxioURI filePath, int replication) {
-    LoadJobAttempt jobAttempt = new LoadJobAttempt(mClient, new
-        LoadConfig(filePath.getPath(), replication), new CountingRetry(3));
-
-    jobAttempt.run();
-
-    return jobAttempt;
-  }
-
-  /**
-   * Add one job.
-   */
-  private void addJob(URIStatus status, int replication) {
-    AlluxioURI filePath = new AlluxioURI(status.getPath());
-    if (status.getInAlluxioPercentage() == 100) {
-      // The file has already been fully loaded into Alluxio.
-      System.out.println(filePath + " is already fully loaded in Alluxio");
-      return;
-    }
-    if (mSubmittedJobAttempts.size() >= mActiveJobs) {
-      // Wait one job to complete.
-      waitJob();
-    }
-    System.out.println(filePath + " loading");
-    mSubmittedJobAttempts.add(newJob(filePath, replication));
-  }
-
-  /**
-   * Distributed loads a file or directory in Alluxio space, makes it resident in memory.
-   *
-   * @param filePath The {@link AlluxioURI} path to load into Alluxio memory
-   * @param replication The replication of file to load into Alluxio memory
-   */
-  private void distributedLoad(AlluxioURI filePath, int replication)
-      throws AlluxioException, IOException {
-    load(filePath, replication);
-    // Wait remaining jobs to complete.
-    drain();
-  }
-
-  /**
-   * Loads a file or directory in Alluxio space, makes it resident in memory.
-   *
-   * @param filePath The {@link AlluxioURI} path to load into Alluxio memory
-   * @throws AlluxioException when Alluxio exception occurs
-   * @throws IOException      when non-Alluxio exception occurs
-   */
-  private void load(AlluxioURI filePath, int replication)
-      throws IOException, AlluxioException {
-    ListStatusPOptions options = ListStatusPOptions.newBuilder().setRecursive(true).build();
-    mFileSystem.iterateStatus(filePath, options, uriStatus -> {
-      if (!uriStatus.isFolder()) {
-        addJob(uriStatus, replication);
-      }
-    });
-  }
-
-  private class LoadJobAttempt extends JobAttempt {
-    private LoadConfig mJobConfig;
-
-    LoadJobAttempt(JobMasterClient client, LoadConfig jobConfig, RetryPolicy retryPolicy) {
-      super(client, retryPolicy);
-      mJobConfig = jobConfig;
-    }
-
-    @Override
-    protected JobConfig getJobConfig() {
-      return mJobConfig;
-    }
-
-    @Override
-    protected void logFailedAttempt(JobInfo jobInfo) {
-      System.out.println(String.format("Attempt %d to load %s failed because: %s",
-          mRetryPolicy.getAttemptCount(), mJobConfig.getFilePath(),
-          jobInfo.getErrorMessage()));
-    }
-
-    @Override
-    protected void logFailed() {
-      System.out.println(String.format("Failed to complete loading %s after %d retries.",
-          mJobConfig.getFilePath(), mRetryPolicy.getAttemptCount()));
-    }
-
-    @Override
-    protected void logCompleted() {
-      System.out.println(String.format("Successfully loaded path %s after %d attempts",
-          mJobConfig.getFilePath(), mRetryPolicy.getAttemptCount()));
-    }
   }
 }
